+++ rails/app/controllers/projects_controller.rb
@@ -3,7 +3,9 @@ require 'component_information'
 class ProjectsController < ApplicationController
   include ::ProjectsHelper
   before_action(except: %i[index create new]) { fetch_project(params[:id]) }
-  before_action(except: %i[show index create new reference_components]) { authorize(params[:project_object].editors) }
+  before_action(except: %i[show index create new reference_components changelog_viewed]) do |_action|
+    authorize(params[:project_object].editors)
+  end
 
   # GET /projects/1
   def show # rubocop:disable Metrics/AbcSize