+++ rails/client/app/bundles/kaiju/components/Component/reducers/refreshedComponent.js
@@ -1,0 +1,14 @@
+import { REFRESH_STORE, SELECT_COMPONENT } from '../actions/actions';
+
+const refreshedComponent = (state = null, { type, id }) => {
+  switch (type) {
+    case SELECT_COMPONENT:
+      return null;
+    case REFRESH_STORE:
+      return id;
+    default:
+      return state;
+  }
+};
+
+export default refreshedComponent;