+++ changelog.md
@@ -1,6 +1,28 @@
-## January 11th, 2018
+## January 28th, 2018
+### Added
+- Clicking a selected component while holding ctrl/cmd will select the parent
+  - See [guide](/guide) for a visual example
+
+## January 25th, 2018
+### Added
+- [A help guide](/guide) (Accessible under the help menu dropdown)
+- A keyboard shortcut for duplicating the selected component (ctrl-d/⌘-d)
+
+## January 19th, 2018
 ### Added
 - Ability to duplicate workspaces from the action bar
+- A read-only banner for unauthorized workspaces
+- Checkbox
+- Radio
+- Tabs
+- Status View
+  - Known issue with align top, requires page refresh
+- Footer to the Content Container
+
+### Deprecated
+- Control (Replaced with Radio and Checkbox)
+- No data view (Replaced with Status View)
+- Error view (Replaced with Status View)
 
 ### Fixed
 - The component drop down search is no longer case sensitive